refactor(api): move csv serializer to src/shared
