+++ api/src/prescription/campaign/domain/services/campaign-csv-export-service.js
@@ -1,5 +1,5 @@
 import * as campaignParticipationService from '../../../../../lib/domain/services/campaign-participation-service.js';
-import * as csvSerializer from '../../../../../lib/infrastructure/serializers/csv/csv-serializer.js';
+import * as csvSerializer from '../../../../shared/infrastructure/serializers/csv/csv-serializer.js';
 import { CampaignAssessmentCsvLine } from '../../infrastructure/utils/CampaignAssessmentCsvLine.js';
 
 export { createOneCsvLine };
